Use normal React composition for theming with _app.js in next6
